Conflicts:
	backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/JudicialWarrantApp.java

+++ backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/persistence/repository/AttachmentTypeRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import com.informatique.gov.judicialwarrant.domain.AttachmentType;
 
 @Repository
-public interface AttachmentTypeRepository extends JpaRepository<AttachmentType, Byte>{
+public interface AttachmentTypeRepository extends JpaRepository<AttachmentType, Long>{
 	
 	@Query("select version from AttachmentType at where at.id = :id")
-	Short findVersionById(@Param("id") Byte id);
+	Short findVersionById(@Param("id") Long id);
 
 }
