+++ backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/rest/handler/AttachmentTypeHandler.java
@@ -9,13 +9,13 @@ import com.informatique.gov.judicialwarrant.rest.dto.AttachmentTypeDto;
 public interface AttachmentTypeHandler extends Serializable {
 	ResponseEntity<List<AttachmentTypeDto>> getAll() throws JudicialWarrantException;
 
-	ResponseEntity<AttachmentTypeDto> getById(Byte id, Short etag) throws JudicialWarrantException;
+	ResponseEntity<AttachmentTypeDto> getById(Long id, Short etag) throws JudicialWarrantException;
 
-	ResponseEntity<AttachmentTypeDto> update(AttachmentTypeDto attachmentTypeDto, Byte id, Short etag)
+	ResponseEntity<AttachmentTypeDto> update(AttachmentTypeDto attachmentTypeDto, Long id, Short etag)
 			throws JudicialWarrantException;
 
 	ResponseEntity<AttachmentTypeDto> save(AttachmentTypeDto attachmentTypeDto)
 			throws JudicialWarrantException;
 
-	ResponseEntity<Void> delete(Byte id) throws JudicialWarrantException;
+	ResponseEntity<Void> delete(Long id) throws JudicialWarrantException;
 }
